ai-service/requirements.txt
fastapi
uvicorn
pydantic
pandas
kafka-python
websockets
google-generativeai
python-dotenv
